refactor: 상품 모델 패키지 경로 변경
