+++ README.md
@@ -33,10 +33,10 @@
 🚀 3단계 - 장바구니(담기)
 - [x] 3단계 피드백 반영
   - [x] 상품목록에서 클릭한 상품정보가 상품상세 페이지에 보이도록 수정하기
-- [ ] 데이터
+- [x] 데이터
   - [x] 장바구니 저장소 추가
   - [x] 장바구니 아이템 모델 추가
-  - [ ] 상품 모델 패키지 경로 변경
+  - [x] 상품 모델 패키지 경로 변경
 - [ ] 상품상세 페이지
   - [ ] 장바구니 담기 기능 구현
 - [ ] 장바구니 페이지
