+++ app/src/main/java/nextstep/shoppingcart/ProductDetailActivity.kt
@@ -6,7 +6,7 @@ import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import nextstep.shoppingcart.data.Product
+import nextstep.shoppingcart.data.model.Product
 import nextstep.shoppingcart.ui.screen.ProductDetailScreen
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
