+++ app/src/main/java/nextstep/shoppingcart/ui/screen/ProductDetailScreen.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil3.compose.AsyncImage
 import nextstep.shoppingcart.R
-import nextstep.shoppingcart.data.Product
+import nextstep.shoppingcart.data.model.Product
 import nextstep.shoppingcart.ui.component.BackButtonAppBar
 import nextstep.shoppingcart.ui.component.BottomButton
 import nextstep.shoppingcart.ui.theme.Black
