+++ app/src/main/java/nextstep/shoppingcart/ui/screen/ProductListScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import nextstep.shoppingcart.R
-import nextstep.shoppingcart.data.Product
+import nextstep.shoppingcart.data.model.Product
 import nextstep.shoppingcart.ui.component.ProductLazyGrid
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
